Support multiple consumer attributes per worker
This changes from a class name based argument for the bin to a hash based one. This is because while for now everything happens outside of code, in the future the code running should be able to pass argument to the consumer to only select specific messages. For example when it needs more memory for bigger tasks. So have a couple of big consumers that do everything, and only have the small consumers pick up the small tasks.